update dai rest server jar, clean up unnecessary files, updates to docker image runtime files
